pyproject.toml
[tool.poetry]
name = "hw-13-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
fastapi-mail = "^1.4.1"
redis = "^5.0.5"
fastapi-limiter = "^0.1.6"
ratelimiter = "^1.2.0.post0"
cloudinary = "^1.40.0"
libgravatar = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
